+++ client/src/components/potions/Potions.tsx
@@ -5,6 +5,7 @@
 // Maybe try to make a HoC because I'm sure we will have many lists of items. 
 // Have the potions be little cards with plus and minus signs on them when calculating ingredients.
 import "./potions.css"
+import { Potion as PotionData, PotionCategory } from "../../modules/potion"
 
 function Potions() {
     return (
