--- client/src/modules/ingredients.ts
@@ -1,13 +1,0 @@
-export interface Ingredient {
-    name: string,
-    imageUrl: string,
-    isHardmode: boolean
-}
-
-export const ingredients: Ingredient[] = [
-    {
-        name: "",
-        imageUrl: "",
-        isHardmode: false
-    }
-]
