+++ client/src/modules/material.ts
@@ -1,0 +1,6 @@
+export interface Material {
+    id: number,
+    name: string,
+    imageUrl: string,
+    isHardmode: boolean
+}
