--- client/src/modules/potion-ingredients.ts
@@ -1,11 +1,0 @@
-export interface PotionIngredient {
-    potionId: number,
-    ingredientId: number
-}
-
-export const potionIngredients: PotionIngredient[] = [
-    {
-        potionId: 1,
-        ingredientId: 2
-    }
-]
