+++ client/src/modules/potion.ts
@@ -1,0 +1,19 @@
+export enum PotionCategory {
+    Buff = 1,
+    Utility,
+    Luck,
+    Fishing,
+    SpawnRates,
+    Mining,
+    Flask,
+    Healing,
+    Mana,
+    Misc
+}
+
+export interface Potion {
+    id: number,
+    name: string,
+    imageUrl: string,
+    category: PotionCategory
+}
